Hook up logic to screen and new layout
- Adds new landing screen layouts
- navigates properly to movie fragments

+++ app/src/main/java/greenberg/moviedbshell/models/recentlyreleasedmodels/RecentlyReleasedResponseItem.kt
@@ -34,7 +34,7 @@ data class RecentlyReleasedResponseItem(
 	val popularity: Double? = null,
 
 	@field:SerializedName("vote_average")
-	val voteAverage: Int? = null,
+	val voteAverage: Double? = null,
 
 	@field:SerializedName("id")
 	val id: Int? = null,
